Update development config template for url information
The app config settings changed in Asymmetrik/mean2-starter#77 to include a url key used for creating url information for mailers. This change results in livereload not working in local development unless developers add the app.url information to config/env/development.js.

To assist developers in having the proper setting, this adds the default settings to the config/env/development.template.js.