Test multidatabase configs.
Mongoose's API does not document the use of a callback function with Mongoose#createConnection. The callback works due to the createConnection passing the arguments passed in to Connection#open, which does accept a callback.

Since the API does not document this usage, this test case should provide a canery in case the behavior changes in the future.